chore: refactor to allow Webpack to bundle only what s necessary.
* added define variable for UI custom flavors allowing flavors to remove unecessary code (builder)
* added define variable to disable XML parser (enabled by default to not break anything)

# Conflicts:
#	packages/core/ui/styling/style-scope.ts
