+++ packages/core/global-types.d.ts
@@ -128,6 +128,13 @@ declare namespace NodeJS {
 		__requireOverride?: (name: string, dir: string) => any;
 	}
 }
+declare const __DEV__: string;
+declare const __CSS_PARSER__: string;
+declare const __NS_WEBPACK__: boolean;
+declare const __UI_CUSTOM_FLAVOR__: boolean;
+declare const __UI_XML_PARSER__: boolean;
+declare const __ANDROID__: boolean;
+declare const __IOS__: boolean;
 
 declare function setTimeout(callback: (...args: any[]) => void, ms: number, ...args: any[]): number;
 declare function clearTimeout(timeoutId: number): void;
