+++ packages/core/ui/list-view/list-view-common.ts
@@ -26,7 +26,7 @@ export abstract class ListViewBase extends ContainerView implements ListViewDefi
 	public _defaultTemplate: KeyedTemplate = {
 		key: 'default',
 		createView: () => {
-			if (this.itemTemplate) {
+			if (!__UI_CUSTOM_FLAVOR__ && this.itemTemplate) {
 				return Builder.parse(this.itemTemplate, this);
 			}
 
@@ -195,7 +195,11 @@ export const itemTemplatesProperty = new Property<ListViewBase, string | Array<K
 	name: 'itemTemplates',
 	valueConverter: (value) => {
 		if (typeof value === 'string') {
-			return Builder.parseMultipleTemplates(value, null);
+			if (__UI_XML_PARSER__) {
+				return Builder.parseMultipleTemplates(value, null);
+			} else {
+				return null;
+			}
 		}
 
 		return value;
